Use fixed length `push` for blockheight, clean up mining code path
We now use a fixed 32 bit for the block height in the unlock script of the coinbase transaction. Before we used an optimal sized push, but that mean that the block size changed depending on the height (mostly when changing from mainnet to a smaller testnet) which was a hassle because the tag moved around.

Also finally got a good path though the code that turns the keno cells into a block and checks to see if it is a winner.